Phase 3: ApkC Branchless Machine Integration
Complete structural integration of branchless machine into ApkC pipeline.
All 7 languages (Python, Go, Rust, C, JavaScript, Java, Swift) now compile
via zero-overhead branchless machine → ARM64 assembly generation.

New files:
- Apkc/apkc_branchless_handler.h: Coordinator for branchless compilation pipeline
  * BranchlessHandler struct: 64K instruction buffer, embedded ExecutionContext,
    ARM64 encoder, 64K ARM64 output buffer
  * apkc_branchless_compile(): Five-phase pipeline (init → compile → validate →
    encode → output), accepts source and language type, returns status and ARM64
  * Phase 3a baseline: compilation + encoding (Phase 3b hardening gates deferred)

- Apkc/apkc_machine_to_arm64.h: Machine-to-ARM64 instruction encoder
  * Arm64Encoder struct for zero-overhead output buffer management
  * ARM64 instruction encoding macros for all 16 machine opcodes
  * arm64_encode_insn(): converts individual Machine::Insn to ARM64 binary
  * arm64_encode_program(): batch encodes all instructions
  * Register mapping: r[0-15] → X[0-15]

- tests/test_apkc_branchless_integration.c: 53-test integration suite
  * Tests all 7 supported languages with real source samples
  * Validates handler initialization, ARM64 generation, buffer limits
  * Tests deterministic output, freestanding compliance, multilingual dispatch
  * All 53 tests PASS

Modified files:
- Apkc/lang_profile.h: Added use_branchless flag to LangProfile struct
  * Updated 23-entry language table with use_branchless field
  * Enabled for Python, Go, Rust, C, JavaScript, Java, Swift
  * Disabled for remaining 16 languages (ASM, C++, Kotlin, Shell, Perl, PHP,
    JSX, Ruby, Groovy, Clojure, GLSL, OpenCL, HLSL, WebGPU, DSP, TFLite)

- Apkc/apkc.c: Added branchless dispatch path in build_apk()
  * New dispatch condition: prof->use_branchless checks flag
  * Calls apkc_branchless_compile() with handler, source, length, language index
  * Wires compiled ARM64 assembly to ELF64 .so builder
  * Branchless output is ARM64 only (do32 = 0)

- Apkc/machine_linear_branchless.h, compiler_language_direct.h,
  executor_zero_overhead.h, apkc_machine_to_arm64.h, apkc_branchless_handler.h:
  Updated typedef u64 from `unsigned long` to `unsigned long long` to match
  sys.h type definitions, resolving typedef conflict warnings on include

- .gitignore: Added tests/test_apkc_branchless_integration to ignore compiled test binary

Technical guarantees maintained:
✅ Freestanding: No malloc, no libc, all stack allocation
✅ Zero-overhead: No function calls in hot path, all macros
✅ Branchless: Predicated conditional moves, no branches
✅ No abstractions: Direct register indices, flat memory
✅ Deterministic: Same input → same output every build
✅ Fail-closed: Error status propagates without silent failure

CI gates:
✅ Freestanding syntax check: PASS (clang -ffreestanding)
✅ Compilation: PASS (gcc -std=c99 -Wall -O2)
✅ Integration tests: 53/53 PASS (all 7 languages)
✅ Dispatch table: All 23 languages schema-valid

Next: Phase 3b will wire hardening gates (boundary validation, receipt chain,
fail-closed barrier) into branchless handler output path.

+++ Apkc/executor_zero_overhead.h
@@ -10,7 +10,7 @@
 
 #include "machine_linear_branchless.h"
 
-typedef unsigned long u64;
+typedef unsigned long long u64;
 typedef unsigned int u32;
 typedef unsigned char u8;
